# storage: optimise sampleRing