@beorn7 2023-05-04
The same query micro benchmarks as used in #11687 (allocs not shown, they haven't really changed): 

```
 benchstat bench_3d26faade.txt bench_b0272255b.txt 
name                                                  old time/op    new time/op    delta
RangeQuery/expr=changes(a_one[1d]),steps=1-16            403µs ± 1%     401µs ± 2%     ~     (p=0.280 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=10-16           490µs ± 1%     469µs ± 5%   -4.29%  (p=0.003 n=8+10)
RangeQuery/expr=changes(a_one[1d]),steps=100-16         1.39ms ± 2%    1.15ms ± 1%  -17.51%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=1000-16        9.29ms ± 1%    8.63ms ± 0%   -7.16%  (p=0.000 n=9+9)
RangeQuery/expr=changes(a_ten[1d]),steps=1-16           3.14ms ± 0%    3.13ms ± 1%     ~     (p=0.631 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=10-16          4.03ms ± 2%    3.80ms ± 0%   -5.82%  (p=0.000 n=10+9)
RangeQuery/expr=changes(a_ten[1d]),steps=100-16         13.2ms ± 2%    10.8ms ± 0%  -18.39%  (p=0.000 n=10+9)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-16        91.7ms ± 3%    85.0ms ± 0%   -7.26%  (p=0.000 n=10+9)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-16       31.8ms ± 1%    32.0ms ± 2%     ~     (p=0.529 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-16      40.3ms ± 1%    38.7ms ± 1%   -3.94%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-16      130ms ± 3%     108ms ± 1%  -16.82%  (p=0.000 n=10+9)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-16     920ms ± 2%     849ms ± 1%   -7.71%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_one[1d]),steps=1-16               394µs ± 1%     398µs ± 1%   +0.81%  (p=0.010 n=9+10)
RangeQuery/expr=rate(a_one[1d]),steps=10-16              442µs ± 1%     441µs ± 1%     ~     (p=0.796 n=9+9)
RangeQuery/expr=rate(a_one[1d]),steps=100-16             900µs ± 2%     910µs ± 0%   +1.14%  (p=0.011 n=9+9)
RangeQuery/expr=rate(a_one[1d]),steps=1000-16           5.44ms ± 1%    5.70ms ± 1%   +4.92%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1-16              3.13ms ± 2%    3.12ms ± 1%     ~     (p=0.497 n=9+10)
RangeQuery/expr=rate(a_ten[1d]),steps=10-16             3.58ms ± 1%    3.61ms ± 1%   +0.93%  (p=0.005 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=100-16            8.13ms ± 1%    8.38ms ± 1%   +3.16%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-16           53.6ms ± 1%    56.2ms ± 1%   +4.86%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-16          31.5ms ± 1%    31.4ms ± 1%     ~     (p=0.340 n=9+9)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-16         35.9ms ± 1%    35.9ms ± 0%     ~     (p=0.074 n=9+8)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-16        81.5ms ± 1%    84.0ms ± 1%   +3.07%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-16        530ms ± 1%     558ms ± 1%   +5.22%  (p=0.000 n=10+9)
```

Some ups and some downs. The largest changes are all wins, but not overly convincing overall. (Still, I like this PR in general because the code looks better.)