@beorn7 2023-05-04
I also tried a few other query benchmarks (simple retrieval rather than range selectors), but no overwhelming changes there. Still, the "shape" of the queries could define how impactful this is, and maybe @Stono's setup hits one of those shapes.